falta comentar em condiçoes

+++ include/k_means.h
@@ -1,9 +1,11 @@
+//representação do ponto, que inclui as coordenadas x e y e o cluster a que ele pertence
 struct point {
   float x;
   float y;
-  int cluster;
+  short int cluster;
 };
 
+//representação do cluster, que inclui as coordenadas x e y
 struct cluster {
   float x;
   float y;
